cleanup and selection
